feat: 게시글 삭제는 해당 게시글을 작성한 사용자만 삭제할 수 있도록 변경

+++ anonymous-board/src/main/java/com/example/anonymousboard/post/service/PostService.java
@@ -71,8 +71,9 @@ public class PostService {
     }
 
     @Transactional
-    public void deletePostById(final Long postId) {
+    public void deletePostById(final AuthInfo authInfo, final Long postId) {
         Post post = findPostObject(postId);
+        validateOwner(authInfo, post);
         postRepository.delete(post);
     }
